# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Magic The Gathering Card API
  description: API for managing Magic The Gathering cards.
  version: 1.0.0
servers:
  - url: http://api.example.com
paths:
  /card:
    post:
      summary: Insert a single Magic The Gathering card into the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestInsertCard'
      responses:
        '200':
          description: Card inserted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseInsertCard'
        '400':
          description: Bad request. Invalid payload format or missing required fields.
        '500':
          description: Internal server error. Failed to insert card into the database.
  /cards:
    post:
      summary: Insert multiple Magic The Gathering cards into the database.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cards:
                  type: string
                  format: binary
      responses:
        '200':
          description: Cards inserted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseConciliateJob'
        '500':
          description: Internal server error. Failed to insert cards into the database.
    get:
      summary: Get Magic The Gathering cards based on filters with pagination.
      parameters:
        - name: set_name
          in: query
          required: false
          description: Filter cards by set name.
          schema:
            type: string
        - name: name
          in: query
          required: false
          description: Filter cards by card name.
          schema:
            type: string
        - name: collector_number
          in: query
          required: false
          description: Filter cards by collector number.
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Page number for pagination (default is 1).
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page (default is 10, max is 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Cards retrieved successfully with pagination information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePaginatedCards'
        '400':
          description: Bad request. Invalid pagination parameters.
        '500':
          description: Internal server error. Failed to retrieve cards.
  /card/{id}:
    get:
      summary: Get a Magic The Gathering card by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the card to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: Card retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCard'
        '400':
          description: Bad request. Invalid card ID format.
        '404':
          description: Card not found.
        '500':
          description: Internal server error. Failed to retrieve the card.
    delete:
      summary: Delete a Magic The Gathering card by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the card to delete.
          schema:
            type: string
      responses:
        '200':
          description: Card deleted successfully.
        '400':
          description: Bad request. Invalid card ID format.
        '500':
          description: Internal server error. Failed to delete the card.
    patch:
      summary: Update a Magic The Gathering card by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the card to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdateCard'
      responses:
        '200':
          description: Card updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCard'
        '400':
          description: Bad request. Invalid payload format or missing required fields.
        '404':
          description: Card not found.
        '500':
          description: Internal server error. Failed to update the card.
  /card-history/{id}:
    get:
      summary: Get the price history of a Magic The Gathering card by its ID with pagination.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the card to get the price history.
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Page number for pagination (default is 1).
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page (default is 10, max is 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Price history retrieved successfully with pagination information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePaginatedCards'
        '400':
          description: Bad request. Invalid card ID format or pagination parameters.
        '404':
          description: Card not found.
        '500':
          description: Internal server error. Failed to get the price history.
  /collection-stats:
    get:
      summary: Get collection statistics including total cards, foil cards, unique sets, and total value.
      responses:
        '200':
          description: Collection statistics retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCollectionStats'
        '500':
          description: Internal server error. Failed to get collection statistics.
components:
  schemas:
    RequestInsertCard:
      type: object
      properties:
        name:
          type: string
        set_name:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
    RequestUpdateCard:
      type: object
      properties:
        name:
          type: string
    ResponseInsertCard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        set:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
    ResponseCard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        set:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
        last_price:
          type: number
        old_price:
          type: number
        price_change:
          type: number
        last_update:
          type: string
          format: date-time
    ResponseConciliateJob:
      type: object
      properties:
        processed:
          type: integer
        not_processed:
          type: integer
    ResponsePaginatedCards:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/ResponseCard'
        total:
          type: integer
          description: Total number of cards matching the filters.
        page:
          type: integer
          description: Current page number.
        limit:
          type: integer
          description: Number of items per page.
        total_pages:
          type: integer
          description: Total number of pages available.
    ResponseCollectionStats:
      type: object
      properties:
        total_cards:
          type: integer
          description: Total number of cards in the collection.
        foil_cards:
          type: integer
          description: Number of foil cards in the collection.
        unique_sets:
          type: integer
          description: Number of unique sets in the collection.
        total_value:
          type: number
          description: Total value of the collection.